implementing material design
WIP
