+++ app/src/main/java/com/ofamosoron/dividimos/ui/composables/home/Home.kt
@@ -35,100 +35,7 @@ fun Home(
     viewModel: MainViewModel = hiltViewModel(),
     navController: NavController,
 ) {
-    val state = viewModel.mainState.collectAsState()
 
-    if (state.value.showAlert) {
-        val context = LocalContext.current.applicationContext
-        Toast.makeText(
-            context,
-            "Convidado já está dividindo esse item",
-            Toast.LENGTH_SHORT
-        ).show()
-        viewModel.onEvent(HomeScreenEvent.ClearAlert)
-    }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(color = MaterialTheme.colorScheme.surface)
-    ) {
-        ChooseDialog(
-            dialogType = state.value.openDialog,
-            action = { viewModel.onEvent(it) }
-        )
-
-        Column(modifier = Modifier.fillMaxSize()) {
-            Header(
-                addNewDish = { navController.navigate(Route.NewDishScreen.url) },
-                addNewGuest = { navController.navigate(Route.NewGuestScreen.url) },
-                actionMenuOptionOneClick = { viewModel.onEvent(HomeScreenEvent.OpenDialog(dialogType = DialogType.ClearTableDialog())) },
-                actionMenuOptionTwoClick = { /* TODO */ },
-                total = state.value.dishes.sumOf { it.price.cents * it.qnt }
-            )
-
-            LongPressDraggable(
-                modifier = Modifier.fillMaxSize()
-            ) {
-                if (state.value.dishes.isEmpty()) {
-                    EmptyScreen(msg = R.string.dishes_alert, icon = R.drawable.ic_empty_table)
-                } else {
-                    Box(
-                        modifier = Modifier.align(Alignment.TopCenter).padding(bottom = 56.dp)
-                    ) {
-                        DishesContainer(
-                            dishes = state.value.dishes,
-                            guests = state.value.dishesToGuests,
-                            onIncreaseClick = { dishUuid ->
-                                viewModel.onEvent(
-                                    HomeScreenEvent.IncreaseDishQuantity(
-                                        dishUuid = dishUuid
-                                    )
-                                )
-                            },
-                            onDecreaseClick = { dishUuid ->
-                                viewModel.onEvent(
-                                    HomeScreenEvent.DecreaseDishQuantity(
-                                        dishUuid = dishUuid
-                                    )
-                                )
-                            }, onDrop = { guestUuid, dishUuid ->
-                                viewModel.onEvent(
-                                    HomeScreenEvent.OnDrop(
-                                        guestUuid = guestUuid,
-                                        dishUuid = dishUuid
-                                    )
-                                )
-                            }, onEditClick = { dishUuid: String ->
-                                navController.navigate(Route.EditDishScreen.withArgs(dishUuid))
-                            })
-                    }
-                }
-
-                if (state.value.guests.isEmpty()) {
-                    Text(
-                        text = stringResource(R.string.guest_alert),
-                        color = MaterialTheme.colorScheme.primary,
-                        style = MaterialTheme.typography.headlineSmall,
-                        modifier = Modifier
-                            .alpha(ALPHA)
-                            .align(Alignment.BottomCenter)
-                            .padding(4.dp)
-                    )
-                } else {
-                    GuestsContainer(
-                        guests = state.value.guests,
-                        cardClick = {
-                            viewModel.onEvent(
-                                HomeScreenEvent.OpenDialog(
-                                    dialogType = DialogType.CheckDialog(guestId = it)
-                                )
-                            )
-                        }
-                    )
-                }
-            }
-        }
-    }
 }
 
 @Composable
